requirements.txt
pyspark
kafka-python
mysql-connector-python
python-dotenv